requirements.txt
asgiref==3.8.1
Django==5.1.7
django-crispy-forms==2.3
joblib==1.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.1
